+++ src/FuncVersion.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IActionContext, IAzureQuickPickItem, IAzureQuickPickOptions } from '@microsoft/vscode-azext-utils';
+import { learnMoreQp } from './constants';
 import { localize } from './localize';
 import { openUrl } from './utils/openUrl';
 
@@ -15,6 +16,7 @@ export enum FuncVersion {
 }
 
 export const latestGAVersion: FuncVersion = FuncVersion.v4;
+export const funcVersionLink: string = 'https://aka.ms/AA1tpij';
 
 export async function promptForFuncVersion(context: IActionContext, message?: string): Promise<FuncVersion> {
     const recommended: string = localize('recommended', '(Recommended)');
@@ -27,14 +29,14 @@ export async function promptForFuncVersion(context: IActionContext, message?: st
 
     picks = picks.filter(p => osSupportsVersion(p.data));
 
-    picks.push({ label: localize('learnMore', '$(link-external) Learn more...'), description: '', data: undefined });
+    picks.push(learnMoreQp);
 
     const options: IAzureQuickPickOptions = { placeHolder: message || localize('selectVersion', 'Select a version'), stepName: 'funcVersion', suppressPersistence: true };
     // eslint-disable-next-line no-constant-condition
     while (true) {
         const version: FuncVersion | undefined = (await context.ui.showQuickPick(picks, options)).data;
         if (version === undefined) {
-            await openUrl('https://aka.ms/AA1tpij');
+            await openUrl(funcVersionLink);
         } else {
             return version;
         }