+++ src/funcCoreTools/installFuncCoreTools.ts
@@ -12,6 +12,8 @@ import { cpUtils } from '../utils/cpUtils';
 import { getBrewPackageName } from './getBrewPackageName';
 import { getNpmDistTag, INpmDistTag } from './getNpmDistTag';
 
+export let lastCoreToolsInstallCommand: string[] = [];
+
 export async function installFuncCoreTools(context: IActionContext, packageManagers: PackageManager[], version?: FuncVersion): Promise<void> {
     version = version || await promptForFuncVersion(context, localize('selectVersion', 'Select the version of the runtime to install'));
 
@@ -20,12 +22,15 @@ export async function installFuncCoreTools(context: IActionContext, packageManag
     switch (packageManagers[0]) {
         case PackageManager.npm:
             const distTag: INpmDistTag = await getNpmDistTag(context, version);
-            await cpUtils.executeCommand(ext.outputChannel, undefined, 'npm', 'install', '-g', `${funcPackageName}@${distTag.tag}`);
+            lastCoreToolsInstallCommand = ['npm', 'install', '-g', `${funcPackageName}@${distTag.tag}`];
+            await cpUtils.executeCommand(ext.outputChannel, undefined, lastCoreToolsInstallCommand[0], ...lastCoreToolsInstallCommand.slice(1));
             break;
         case PackageManager.brew:
             const brewPackageName: string = getBrewPackageName(version);
-            await cpUtils.executeCommand(ext.outputChannel, undefined, 'brew', 'tap', 'azure/functions');
-            await cpUtils.executeCommand(ext.outputChannel, undefined, 'brew', 'install', brewPackageName);
+            lastCoreToolsInstallCommand = ['brew', 'tap', 'azure/functions'];
+            await cpUtils.executeCommand(ext.outputChannel, undefined, lastCoreToolsInstallCommand[0], ...lastCoreToolsInstallCommand.slice(1));
+            lastCoreToolsInstallCommand = ['brew', 'install', brewPackageName];
+            await cpUtils.executeCommand(ext.outputChannel, undefined, lastCoreToolsInstallCommand[0], ...lastCoreToolsInstallCommand.slice(1));
             break;
         default:
             throw new RangeError(localize('invalidPackageManager', 'Invalid package manager "{0}".', packageManagers[0]));